PR-13: extract control/server.rs and control/persistent.rs
Behaviour-preserving relocation. ~545 LOC moved out of main.rs.
- control/server.rs: DbusWindowFocusService + its zbus #[interface]
  impl (kept colocated — macro requires same file),
  resolve_runtime_unpause_context, DbusServiceRegistration + Drop,
  register_dbus_service variants.
- control/persistent.rs: DBUS_RECONNECT_DELAYS_MS,
  dbus_reconnect_delay, wait_for_dbus_reconnect_retry,
  PersistentDbusServiceGuard + Drop, start/run persistent
  service variants.
- control/mod.rs: pub(crate) mod server / persistent and
  pub(crate) use {server::*, persistent::*} for test reach.

main.rs: 2501 -> 1979 LOC. All consumers of the moved items lived
in main.rs already; no cross-call-site updates needed.

Verification: cargo build clean; cargo test --bin kanata-switcher 261/0.
