+++ src/daemon/control/mod.rs
@@ -1,4 +1,8 @@
 pub(crate) mod client;
+pub(crate) mod server;
+pub(crate) mod persistent;
+pub(crate) use server::*;
+pub(crate) use persistent::*;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) enum ControlCommand {
