+++ src/daemon/main.rs
@@ -21,7 +21,6 @@ use std::time::Duration;
 use tokio::sync::{Mutex as TokioMutex, mpsc, oneshot, watch};
 use uuid::Uuid;
 use zbus::Connection;
-use zbus::object_server::SignalEmitter;
 use zbus::zvariant::{OwnedObjectPath, OwnedValue, Str, Structure, Value};
 
 mod constants;
@@ -52,7 +51,7 @@ use args::*;
 use autostart::*;
 use broadcasters::*;
 use kanata::*;
-use control::{ControlCommand, ControlDispatch};
+use control::*;
 use control::client::*;
 use pause::*;
 use focus_pipeline::*;
@@ -66,7 +65,7 @@ use backends::wayland::*;
 
 #[cfg(test)]
 #[allow(unused_imports)]
-pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*, broadcasters::*, kanata::*, control::*, control::client::*, pause::*, focus_pipeline::*, lifecycle::*, lifecycle::logind::*, lifecycle::startup::*, display_override::*, supervisor::*, supervisor::capabilities::*, backends::*, backends::gnome::*, backends::kde::*, backends::kde::script::*, backends::kde::probe::*, backends::wayland::*};
+pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*, broadcasters::*, kanata::*, control::*, control::client::*, control::server::*, control::persistent::*, pause::*, focus_pipeline::*, lifecycle::*, lifecycle::logind::*, lifecycle::startup::*, display_override::*, supervisor::*, supervisor::capabilities::*, backends::*, backends::gnome::*, backends::kde::*, backends::kde::script::*, backends::kde::probe::*, backends::wayland::*};
 
 
 // === SNI Indicator ===
@@ -1790,527 +1789,6 @@ fn setup_gnome_extension(auto_install: bool) {
         println!("[GNOME] Please restart GNOME Shell to activate the extension.");
         println!("[GNOME]   - Press Alt+F2, type \"r\", press Enter (X11 only)");
         println!("[GNOME]   - Or log out and log back in (Wayland)");
-    }
-}
-
-// === DBus Backend (shared by GNOME and KDE) ===
-
-#[derive(Debug)]
-struct DbusWindowFocusService {
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    runtime_handle: tokio::runtime::Handle,
-    status_broadcaster: StatusBroadcaster,
-    restart_handle: RestartHandle,
-    pause_broadcaster: PauseBroadcaster,
-    env: Environment,
-    focus_query_connection: Connection,
-    is_kde6: bool,
-    runtime_environment: Option<RuntimeEnvironmentBroadcaster>,
-}
-
-#[zbus::interface(name = "com.github.kanata.Switcher")]
-impl DbusWindowFocusService {
-    async fn window_focus(&self, window_class: &str, window_title: &str) {
-        let win = WindowInfo {
-            class: window_class.to_string(),
-            title: window_title.to_string(),
-            is_native_terminal: false,
-        };
-
-        if self.pause_broadcaster.is_paused() {
-            return;
-        }
-
-        let default_layer = self
-            .runtime_handle
-            .block_on(async { self.kanata.default_layer().await })
-            .unwrap_or_default();
-
-        let actions = self.runtime_handle.block_on(async {
-            update_status_for_focus(
-                &self.handler,
-                &self.status_broadcaster,
-                &win,
-                &self.kanata,
-                &default_layer,
-            )
-            .await
-        });
-
-        if let Some(actions) = actions {
-            let kanata = self.kanata.clone();
-            self.runtime_handle
-                .block_on(async { execute_focus_actions(&kanata, actions).await });
-        }
-    }
-
-    async fn get_status(&self) -> (String, Vec<String>, String) {
-        let snapshot = self.status_broadcaster.snapshot();
-        (
-            snapshot.layer,
-            snapshot.virtual_keys,
-            snapshot.layer_source.as_str().to_string(),
-        )
-    }
-
-    async fn get_paused(&self) -> bool {
-        self.pause_broadcaster.is_paused()
-    }
-
-    #[zbus(signal)]
-    async fn status_changed(
-        signal_emitter: &SignalEmitter<'_>,
-        layer: &str,
-        virtual_keys: &[&str],
-        source: &str,
-    ) -> zbus::Result<()>;
-
-    #[zbus(signal)]
-    async fn paused_changed(signal_emitter: &SignalEmitter<'_>, paused: bool) -> zbus::Result<()>;
-
-    async fn restart(&self) {
-        println!("[Restart] Restart requested via DBus");
-        self.restart_handle.request();
-    }
-
-    async fn pause(&self) {
-        pause_daemon(
-            &self.pause_broadcaster,
-            &self.handler,
-            &self.status_broadcaster,
-            &self.kanata,
-            &self.runtime_handle,
-            "via DBus",
-        );
-    }
-
-    async fn unpause(&self) {
-        let context = match &self.runtime_environment {
-            Some(runtime_environment) => resolve_runtime_unpause_context(runtime_environment).await,
-            None => UnpauseContext {
-                env: self.env,
-                connection: Some(self.focus_query_connection.clone()),
-                is_kde6: self.is_kde6,
-            },
-        };
-        unpause_daemon(
-            context.env,
-            context.connection,
-            context.is_kde6,
-            &self.pause_broadcaster,
-            &self.handler,
-            &self.status_broadcaster,
-            &self.kanata,
-            &self.runtime_handle,
-            "via DBus",
-        );
-    }
-}
-
-async fn resolve_runtime_unpause_context(
-    runtime_environment: &RuntimeEnvironmentBroadcaster,
-) -> UnpauseContext {
-    let env = runtime_environment.current();
-    let connection = if environment_requires_focus_query_connection(env) {
-        Some(Connection::session().await.unwrap_or_else(|error| {
-            panic!(
-                "[DBus] Failed to connect to session bus for unpause focus query: {}",
-                error
-            )
-        }))
-    } else {
-        None
-    };
-    let is_kde6 = if env == Environment::Kde {
-        let connection_ref = connection
-            .as_ref()
-            .expect("KDE runtime unpause context requires session connection");
-        resolve_kde_runtime_query_mode_with_retry(connection_ref)
-            .await
-            .unwrap_or_else(|error| {
-                panic!(
-                    "[KDE] Failed to resolve runtime query mode for unpause: {}",
-                    error
-                )
-            })
-    } else {
-        false
-    };
-    UnpauseContext {
-        env,
-        connection,
-        is_kde6,
-    }
-}
-
-struct DbusServiceRegistration {
-    _connection: Connection,
-    status_signal_task: tokio::task::JoinHandle<()>,
-    pause_signal_task: tokio::task::JoinHandle<()>,
-}
-
-impl Drop for DbusServiceRegistration {
-    fn drop(&mut self) {
-        self.status_signal_task.abort();
-        self.pause_signal_task.abort();
-    }
-}
-
-#[cfg(test)]
-async fn register_dbus_service(
-    connection: &Connection,
-    focus_query_connection: Connection,
-    env: Environment,
-    is_kde6: bool,
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    restart_handle: RestartHandle,
-    pause_broadcaster: PauseBroadcaster,
-    effective_name: &str,
-) -> Result<DbusServiceRegistration, DynError> {
-    register_dbus_service_with_runtime_environment(
-        connection,
-        focus_query_connection,
-        env,
-        is_kde6,
-        kanata,
-        handler,
-        status_broadcaster,
-        restart_handle,
-        pause_broadcaster,
-        None,
-        effective_name,
-    )
-    .await
-}
-
-async fn register_dbus_service_with_runtime_environment(
-    connection: &Connection,
-    focus_query_connection: Connection,
-    env: Environment,
-    is_kde6: bool,
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    restart_handle: RestartHandle,
-    pause_broadcaster: PauseBroadcaster,
-    runtime_environment: Option<RuntimeEnvironmentBroadcaster>,
-    effective_name: &str,
-) -> Result<DbusServiceRegistration, DynError> {
-    let service = DbusWindowFocusService {
-        kanata,
-        handler,
-        runtime_handle: tokio::runtime::Handle::current(),
-        status_broadcaster: status_broadcaster.clone(),
-        restart_handle,
-        pause_broadcaster: pause_broadcaster.clone(),
-        env,
-        focus_query_connection,
-        is_kde6,
-        runtime_environment,
-    };
-
-    connection.object_server().at(DBUS_PATH, service).await?;
-
-    connection.request_name(effective_name).await?;
-
-    let mut receiver = status_broadcaster.subscribe();
-    let signal_emitter = SignalEmitter::new(connection, DBUS_PATH)?.into_owned();
-    let initial_status = status_broadcaster.snapshot();
-    let initial_virtual_keys: Vec<&str> = initial_status
-        .virtual_keys
-        .iter()
-        .map(|vk| vk.as_str())
-        .collect();
-    DbusWindowFocusService::status_changed(
-        &signal_emitter,
-        &initial_status.layer,
-        &initial_virtual_keys,
-        initial_status.layer_source.as_str(),
-    )
-    .await?;
-    let signal_emitter_task = signal_emitter.clone();
-    let status_signal_task = tokio::spawn(async move {
-        let mut last = receiver.borrow().clone();
-        loop {
-            if receiver.changed().await.is_err() {
-                break;
-            }
-            let current = receiver.borrow().clone();
-            if current != last {
-                let virtual_keys: Vec<&str> =
-                    current.virtual_keys.iter().map(|vk| vk.as_str()).collect();
-                let _ = DbusWindowFocusService::status_changed(
-                    &signal_emitter_task,
-                    &current.layer,
-                    &virtual_keys,
-                    current.layer_source.as_str(),
-                )
-                .await;
-                last = current;
-            }
-        }
-    });
-
-    let mut pause_receiver = pause_broadcaster.subscribe();
-    let pause_emitter = signal_emitter.clone();
-    DbusWindowFocusService::paused_changed(&pause_emitter, pause_broadcaster.is_paused()).await?;
-    let pause_signal_task = tokio::spawn(async move {
-        let mut last = *pause_receiver.borrow();
-        loop {
-            if pause_receiver.changed().await.is_err() {
-                break;
-            }
-            let current = *pause_receiver.borrow();
-            if current != last {
-                let _ = DbusWindowFocusService::paused_changed(&pause_emitter, current).await;
-                last = current;
-            }
-        }
-    });
-
-    Ok(DbusServiceRegistration {
-        _connection: connection.clone(),
-        status_signal_task,
-        pause_signal_task,
-    })
-}
-
-const DBUS_RECONNECT_DELAYS_MS: &[u64] = &[250, 1000, 2000];
-
-fn dbus_reconnect_delay(attempt: usize) -> Duration {
-    let index = attempt.min(DBUS_RECONNECT_DELAYS_MS.len() - 1);
-    Duration::from_millis(DBUS_RECONNECT_DELAYS_MS[index])
-}
-
-async fn wait_for_dbus_reconnect_retry(
-    reconnect_attempt: &mut usize,
-    shutdown_receiver: &mut watch::Receiver<bool>,
-    restart_receiver: &mut watch::Receiver<bool>,
-    prefix: &str,
-    error: String,
-) {
-    let delay = dbus_reconnect_delay(*reconnect_attempt);
-    eprintln!("{}; retrying in {}ms: {}", prefix, delay.as_millis(), error);
-    tokio::select! {
-        _ = tokio::time::sleep(delay) => {}
-        _ = shutdown_receiver.changed() => {}
-        _ = restart_receiver.changed() => {}
-    }
-    *reconnect_attempt += 1;
-}
-
-struct PersistentDbusServiceGuard {
-    task: tokio::task::JoinHandle<()>,
-}
-
-impl Drop for PersistentDbusServiceGuard {
-    fn drop(&mut self) {
-        self.task.abort();
-    }
-}
-
-fn start_persistent_dbus_service(
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    restart_handle: RestartHandle,
-    pause_broadcaster: PauseBroadcaster,
-    runtime_environment: RuntimeEnvironmentBroadcaster,
-    shutdown_handle: ShutdownHandle,
-    effective_name: String,
-) -> PersistentDbusServiceGuard {
-    start_persistent_dbus_service_with_connector(
-        || async {
-            Connection::session()
-                .await
-                .map_err(|error| -> DynError { Box::new(error) })
-        },
-        kanata,
-        handler,
-        status_broadcaster,
-        restart_handle,
-        pause_broadcaster,
-        runtime_environment,
-        shutdown_handle,
-        effective_name,
-    )
-}
-
-fn start_persistent_dbus_service_with_connector<C, CFut>(
-    connector: C,
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    restart_handle: RestartHandle,
-    pause_broadcaster: PauseBroadcaster,
-    runtime_environment: RuntimeEnvironmentBroadcaster,
-    shutdown_handle: ShutdownHandle,
-    effective_name: String,
-) -> PersistentDbusServiceGuard
-where
-    C: Fn() -> CFut + Send + Sync + 'static,
-    CFut: std::future::Future<Output = Result<Connection, DynError>> + Send + 'static,
-{
-    let task = tokio::spawn(async move {
-        run_persistent_dbus_service_with_connector(
-            connector,
-            kanata,
-            handler,
-            status_broadcaster,
-            restart_handle,
-            pause_broadcaster,
-            runtime_environment,
-            shutdown_handle,
-            effective_name,
-        )
-        .await;
-    });
-    PersistentDbusServiceGuard { task }
-}
-
-async fn run_persistent_dbus_service_with_connector<C, CFut>(
-    connector: C,
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    restart_handle: RestartHandle,
-    pause_broadcaster: PauseBroadcaster,
-    runtime_environment: RuntimeEnvironmentBroadcaster,
-    shutdown_handle: ShutdownHandle,
-    effective_name: String,
-) where
-    C: Fn() -> CFut + Send + Sync + 'static,
-    CFut: std::future::Future<Output = Result<Connection, DynError>> + Send + 'static,
-{
-    let mut restart_receiver = restart_handle.subscribe();
-    let mut shutdown_receiver = shutdown_handle.subscribe();
-    let mut reconnect_attempt = 0usize;
-
-    loop {
-        if *shutdown_receiver.borrow() || *restart_receiver.borrow() {
-            return;
-        }
-
-        let connection = match connector().await {
-            Ok(connection) => {
-                reconnect_attempt = 0;
-                connection
-            }
-            Err(error) => {
-                wait_for_dbus_reconnect_retry(
-                    &mut reconnect_attempt,
-                    &mut shutdown_receiver,
-                    &mut restart_receiver,
-                    "[DBus] Session bus unavailable",
-                    error.to_string(),
-                )
-                .await;
-                continue;
-            }
-        };
-
-        let registration = match register_dbus_service_with_runtime_environment(
-            &connection,
-            connection.clone(),
-            Environment::Unknown,
-            false,
-            kanata.clone(),
-            handler.clone(),
-            status_broadcaster.clone(),
-            restart_handle.clone(),
-            pause_broadcaster.clone(),
-            Some(runtime_environment.clone()),
-            &effective_name,
-        )
-        .await
-        {
-            Ok(registration) => {
-                println!("[DBus] Control service registered as {}", effective_name);
-                reconnect_attempt = 0;
-                registration
-            }
-            Err(error) => {
-                wait_for_dbus_reconnect_retry(
-                    &mut reconnect_attempt,
-                    &mut shutdown_receiver,
-                    &mut restart_receiver,
-                    "[DBus] Failed to register control service",
-                    error.to_string(),
-                )
-                .await;
-                continue;
-            }
-        };
-
-        let proxy = match zbus::fdo::DBusProxy::new(&connection).await {
-            Ok(proxy) => proxy,
-            Err(error) => {
-                drop(registration);
-                wait_for_dbus_reconnect_retry(
-                    &mut reconnect_attempt,
-                    &mut shutdown_receiver,
-                    &mut restart_receiver,
-                    "[DBus] Failed to create DBus proxy for name-loss monitoring",
-                    error.to_string(),
-                )
-                .await;
-                continue;
-            }
-        };
-        let mut name_lost = match proxy
-            .receive_name_lost_with_args(&[(0, effective_name.as_str())])
-            .await
-        {
-            Ok(stream) => stream,
-            Err(error) => {
-                drop(registration);
-                wait_for_dbus_reconnect_retry(
-                    &mut reconnect_attempt,
-                    &mut shutdown_receiver,
-                    &mut restart_receiver,
-                    "[DBus] Failed to subscribe to NameLost",
-                    error.to_string(),
-                )
-                .await;
-                continue;
-            }
-        };
-
-        tokio::select! {
-            _ = shutdown_receiver.changed() => {
-                drop(registration);
-            }
-            _ = restart_receiver.changed() => {
-                drop(registration);
-            }
-            signal = name_lost.next() => {
-                match signal {
-                    Some(signal) => {
-                        match signal.args() {
-                            Ok(args) => {
-                                eprintln!(
-                                    "[DBus] Lost well-known name {}; re-registering",
-                                    args.name()
-                                );
-                            }
-                            Err(error) => {
-                                eprintln!(
-                                    "[DBus] Failed to decode NameLost signal; re-registering: {}",
-                                    error
-                                );
-                            }
-                        }
-                    }
-                    None => {
-                        eprintln!("[DBus] NameLost stream terminated; re-registering");
-                    }
-                }
-                drop(registration);
-            }
-        }
     }
 }
